+++ sites/public/src/components/applications/ApplicationMultiselectQuestionStep.tsx
@@ -9,9 +9,9 @@ import {
   AppearanceStyleType,
   ProgressNav,
 } from "@bloom-housing/ui-components"
-import FormsLayout from "../layouts/forms"
-import FormBackLink from "./forms/applications/FormBackLink"
-import { useFormConductor } from "../lib/hooks"
+import FormsLayout from "../../layouts/forms"
+import FormBackLink from "./FormBackLink"
+import { useFormConductor } from "../../lib/hooks"
 import { ApplicationSection, MultiselectOption } from "@bloom-housing/backend-core/types"
 import {
   OnClientSide,
@@ -29,7 +29,7 @@ import {
   mapRadiosToApi,
   listingSectionQuestions,
 } from "@bloom-housing/shared-helpers"
-import { UserStatus } from "../lib/constants"
+import { UserStatus } from "../../lib/constants"
 
 export interface ApplicationMultiselectQuestionStepProps {
   applicationSection: ApplicationSection
