+++ sites/public/src/components/applications/ApplicationTimeout.ts
@@ -2,7 +2,7 @@ import { createElement, useContext } from "react"
 import { t } from "@bloom-housing/ui-components"
 import { AuthContext, IdleTimeout } from "@bloom-housing/shared-helpers"
 
-import { AppSubmissionContext } from "../../../lib/AppSubmissionContext"
+import { AppSubmissionContext } from "../../lib/applications/AppSubmissionContext"
 
 const ApplicationTimeout = () => {
   const { profile } = useContext(AuthContext)
