+++ sites/public/src/components/listing/ListingView.tsx
@@ -56,9 +56,9 @@ import {
   getSummariesTable,
 } from "@bloom-housing/shared-helpers"
 import dayjs from "dayjs"
-import { ErrorPage } from "../pages/_error"
-import { useGetApplicationStatusProps } from "../lib/hooks"
-import { getGenericAddress, openInFuture } from "../lib/helpers"
+import { ErrorPage } from "../../pages/_error"
+import { useGetApplicationStatusProps } from "../../lib/hooks"
+import { getGenericAddress, openInFuture } from "../../lib/helpers"
 
 interface ListingProps {
   listing: Listing
