+++ sites/public/src/pages/account/dashboard.tsx
@@ -12,7 +12,7 @@ import {
 } from "@bloom-housing/ui-components"
 import { PageView, pushGtmEvent, AuthContext, RequireLogin } from "@bloom-housing/shared-helpers"
 import Layout from "../../layouts/application"
-import { MetaTags } from "../../components/MetaTags"
+import { MetaTags } from "../../components/shared/MetaTags"
 import { UserStatus } from "../../lib/constants"
 
 interface DashboardProps {
