+++ sites/public/src/pages/applications/contact/alternate-contact-contact.tsx
@@ -24,7 +24,7 @@ import {
   stateKeys,
   AuthContext,
 } from "@bloom-housing/shared-helpers"
-import FormBackLink from "../../../components/forms/applications/FormBackLink"
+import FormBackLink from "../../../components/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 import { useContext, useEffect } from "react"
 import { UserStatus } from "../../../lib/constants"
