+++ sites/public/src/pages/applications/household/member.tsx
@@ -20,7 +20,7 @@ import {
 import { HouseholdMember, Member } from "@bloom-housing/backend-core/types"
 import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
-import { AppSubmissionContext } from "../../../lib/AppSubmissionContext"
+import { AppSubmissionContext } from "../../../lib/applications/AppSubmissionContext"
 import React, { useContext, useEffect } from "react"
 import { Select } from "@bloom-housing/ui-components/src/forms/Select"
 import {
