+++ sites/public/src/pages/applications/preferences/all.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ApplicationSection } from "@bloom-housing/backend-core/types"
 import { t } from "@bloom-housing/ui-components"
-import ApplicationMultiselectQuestionStep from "../../../components/ApplicationMultiselectQuestionStep"
+import ApplicationMultiselectQuestionStep from "../../../components/applications/ApplicationMultiselectQuestionStep"
 import { listingSectionQuestions } from "@bloom-housing/shared-helpers"
 import { useFormConductor } from "../../../lib/hooks"
