+++ sites/public/src/pages/listing/[id]/[slug].tsx
@@ -11,8 +11,8 @@ import {
 } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../../../lib/constants"
 import Layout from "../../../layouts/application"
-import { ListingView } from "../../../components/ListingView"
-import { MetaTags } from "../../../components/MetaTags"
+import { ListingView } from "../../../components/listing/ListingView"
+import { MetaTags } from "../../../components/shared/MetaTags"
 import { ErrorPage } from "../../_error"
 import dayjs from "dayjs"
 import { fetchJurisdictionByName } from "../../../lib/hooks"
